fixed bug in pipeline
