+++ src/ac_pipeline/config.env
@@ -3,7 +3,7 @@ NODES=/projects/stars/Data_services/biolink3/graphs/Baseline/61f48d345103d107/no
 #/projects/stars/Data_services/translator-ingests/releases/translator_kg/2026_03_27/nodes.jsonl
 EDGES=/projects/stars/Data_services/biolink3/graphs/Baseline/61f48d345103d107/redundant_edges.jsonl
 #/projects/stars/Data_services/translator-ingests/releases/translator_kg/2026_03_27/edges.jsonl
-OUTDIR=/projects/stars/var/answer_coalesce/2026-04-15
+OUTDIR=/projects/stars/var/answer_coalesce/2026-04-21
 #/projects/stars/var/answer_coalesce/2026-04-16-translatorKG
