+++ src/graph_coalescence/build_redis_files.py
@@ -1,24 +1,22 @@
 import csv
 from collections import defaultdict
-import json,jsonlines
+import os
+import json
+import orjson
+import gzip
 import requests
 import bmt
 
-#A link is: link = (other_node, predicate, source)
+try:
+    from tqdm import tqdm
+    TQDM_AVAILABLE = True
+except ImportError:
+    TQDM_AVAILABLE = False
+
+# A link is: link = (other_node, predicate, source)
 # source is a boolean that is true if node is the source, and false if other_node is the source
-
-def str2list(nlabels):
-    x = nlabels[1:-1] #strip [ ]
-    parts = x.split(',')
-    labs = [ pi[1:-1] for pi in parts] #strip ""
-    return labs
-
-def add_labels(lfile,nid,nlabels,done):
-    if nid in done:
-        return
-    labs = str2list(nlabels)
-    lfile.write(f'{nid}\t{labs}\n')
-    done.add(nid)
+FILTER_PREDICATES = ["biolink:related_to_at_concept_level", "biolink:related_to_at_instance_level"]
+BLOCKLIST_URL = "https://raw.githubusercontent.com/NCATSTranslator/Relay/master/config/blocklist.json"
 
 
 def parse_line(line):
@@ -37,25 +35,60 @@ def parse_line(line):
     if pred == 'biolink:expressed_in' and target_id.startswith('UBERON'):
         return source_id, target_id, None, None
     predicate_parts = {'predicate': line['predicate']}
-    for key,value in line.items():
+    for key, value in line.items():
         if 'qualifier' in key:
             predicate_parts[key] = value
-    predicate_string = json.dumps(predicate_parts,sort_keys=True)
+    predicate_string = json.dumps(predicate_parts, sort_keys=True)
     return source_id, target_id, predicate_string, pred
+
+
+def extract_prov(line):
+    # old format: ROBOKOP
+    pks = line.get('primary_knowledge_source')
+    if pks:
+        print ("old")
+        prov = {'primary_knowledge_source': pks}
+        aks = line.get('aggregator_knowledge_source')
+        if aks:
+            prov['aggregator_knowledge_source'] = aks
+        return prov
+
+    print("new")
+    # new format: Translator
+    sources = line.get('sources')
+    if not sources:
+        return {}
+
+    prov = {}
+    for source in sources:
+        role = source.get('resource_role')
+        rid = source.get('resource_id')
+        if not role or not rid:
+            continue
+        if role == 'primary_knowledge_source':
+            prov['primary_knowledge_source'] = rid
+        elif role == 'aggregator_knowledge_source':
+            prov.setdefault('aggregator_knowledge_source', []).append(rid)
+    return prov
+
 
 def get_filter_nodes():
     """Pull the ARS blocklist of nodes that we don't want to return. This consists of a lot of UMLS terms
     that we don't have anyway, but also a bunch of very generic terms (Disease, Human) that are not useful
     and make lots of links in our database."""
 
-    blocklist_url = "https://raw.githubusercontent.com/NCATSTranslator/Relay/master/config/blocklist.json"
-
-    blocklist = json.loads(requests.get(blocklist_url).text)
+    blocklist = json.loads(requests.get(BLOCKLIST_URL).text)
     return set(blocklist)
 
-def go(input_node_file="nodes.jsonl", output_nodelabels='nodelabels.txt', output_nodenames='nodenames.txt',
-       output_category_count='category_count.txt', input_edge_file='edges.jsonl', output_prov='prov.txt',
-       output_links='links.txt', output_backlinks='backlinks.txt'):
+
+def quick_jsonl_file_iterator(json_file, is_gzip=False):
+    with gzip.open(json_file, 'rt') if is_gzip \
+            else open(json_file, 'r', encoding='utf-8') as fp:
+        for line in fp:
+            yield orjson.loads(line)
+
+
+def generate_ac_files(input_node_file, input_edge_file, output_dir):
     """Given a dump of a graph a la robokop, produce 3 files:
     nodelabels.txt which is 2 columns, (id), (list of labels):
     CAID:CA13418922 ['named_thing', 'biological_entity', 'molecular_entity', 'genomic_entity', 'sequence_variant']
@@ -79,33 +112,40 @@ def go(input_node_file="nodes.jsonl", output_nodelabels='nodelabels.txt', output
      True and False (subject) edges into True. Then, in the TRAPI version, we'll need to be careful
      to make sure and look for the right thing, even if the input TRAPI is pointed into a False direction.
     """
+    output_nodelabels_filepath = os.path.join(output_dir, 'nodelabels.txt')
+    output_nodenames_filepath = os.path.join(output_dir, 'nodenames.txt')
+    output_category_count_filepath = os.path.join(output_dir, 'category_count.txt')
+    output_prov_filepath = os.path.join(output_dir, 'prov.txt')
+    output_links_filepath = os.path.join(output_dir, 'links.txt')
+    output_backlinks_filepath = os.path.join(output_dir, 'backlinks.txt')
+
     tk = bmt.Toolkit()
+
     filter_nodes = get_filter_nodes()
-    filter_predicates = ["biolink:related_to_at_concept_level", "biolink:related_to_at_instance_level"]
     categories = {}
     catcount = defaultdict(int)
-    with jsonlines.open(input_node_file,'r') as nodefile, open(output_nodelabels, 'w') as labelfile, open(output_nodenames, 'w') as namefile:
-        for node in nodefile:
-            if node["id"].startswith('CAID'):
+    with open(output_nodelabels_filepath, 'w') as labelfile, open(output_nodenames_filepath, 'w') as namefile:
+        for node in tqdm(quick_jsonl_file_iterator(input_node_file)) if TQDM_AVAILABLE else quick_jsonl_file_iterator(input_node_file):
+            node_id = node["id"]
+            if node_id.startswith('CAID') or node_id in filter_nodes:
                 continue
-            if node["id"] in filter_nodes:
-                continue
-            labelfile.write(f'{node["id"]}\t{node["category"]}\n')
-            categories[node["id"]] = node["category"]
-            for c in node['category']:
+            node_category = node["category"]
+            labelfile.write(f'{node_id}\t{node_category}\n')
+            categories[node_id] = node_category
+            for c in node_category:
                 catcount[c] += 1
-            name = node.get("name","")
+            name = node.get("name", "")
             if name is not None:
-                name = name.encode('ascii',errors='ignore').decode(encoding="utf-8")
-            namefile.write(f'{node["id"]}\t{name}\n')
+                name = name.encode('ascii', errors='ignore').decode(encoding="utf-8")
+            namefile.write(f'{node_id}\t{name}\n')
     nodes_to_links = defaultdict(list)
     edgecounts = defaultdict(int)
-    with open(output_category_count, 'w') as catcountout:
-        for c,v in catcount.items():
+    with open(output_category_count_filepath, 'w') as catcountout:
+        for c, v in catcount.items():
             catcountout.write(f'{c}\t{v}\n')
-    with jsonlines.open(input_edge_file, 'r') as inf, open(output_prov, 'w') as provout:
+    with open(output_prov_filepath, 'w') as provout:
         nl = 0
-        for line in inf:
+        for line in quick_jsonl_file_iterator(input_edge_file):
             if line["subject"].startswith('CAID') or line["object"].startswith('CAID'):
                 continue
             nl += 1
@@ -114,9 +154,9 @@ def go(input_node_file="nodes.jsonl", output_nodelabels='nodelabels.txt', output
                 continue
             if pred is None:
                 continue
-            if just_predicate in filter_predicates:
+            if just_predicate in FILTER_PREDICATES:
                 continue
-            source_link = (target_id,pred,True)
+            source_link = (target_id, pred, True)
             #Here's how we're handling symmetric predicates.
             # The source link and count is going to be just the same, but we're going to modify the target link
             # to look like it's going from target to source.
@@ -124,15 +164,15 @@ def go(input_node_file="nodes.jsonl", output_nodelabels='nodelabels.txt', output
                 target_is_source = True
             else:
                 target_is_source = False
-            target_link = (source_id,pred,target_is_source)
+            target_link = (source_id, pred, target_is_source)
             nodes_to_links[source_id].append(source_link)
             nodes_to_links[target_id].append(target_link)
             for tcategory in set(categories[target_id]):
-                edgecounts[ (source_id, pred, True, tcategory) ] += 1
+                edgecounts[(source_id, pred, True, tcategory)] += 1
             for scategory in set(categories[source_id]):
-                edgecounts[ (target_id, pred, target_is_source, scategory) ] += 1
-            pkey=f'{source_id} {pred} {target_id}'
-            prov = {x:line[x] for x in ['biolink:primary_knowledge_source','biolink:aggregator_knowledge_source'] if x in line}
+                edgecounts[(target_id, pred, target_is_source, scategory)] += 1
+            pkey = f'{source_id} {pred} {target_id}'
+            prov = extract_prov(line)
             if not prov or not pkey:
                 continue
             provout.write(f'{pkey}\t{json.dumps(prov)}\n')
@@ -140,16 +180,16 @@ def go(input_node_file="nodes.jsonl", output_nodelabels='nodelabels.txt', output
                 print(nl)
         print('node labels and names done')
 
-    with open(output_links, 'w') as outf:
-        for node,links in nodes_to_links.items():
+    with open(output_links_filepath, 'w') as outf:
+        for node, links in nodes_to_links.items():
             outf.write(f'{node}\t{json.dumps(links)}\n')
         print('links done')
 
-    with open(output_backlinks, 'w') as outf:
-        for key,value in edgecounts.items():
+    with open(output_backlinks_filepath, 'w') as outf:
+        for key, value in edgecounts.items():
             outf.write(f'{key}\t{value}\n')
         print('backlinks done')
 
 
 if __name__ == '__main__':
-    go()
+    generate_ac_files()
